Move all of mwnicks code into logical files for easier reading and removal later
