+++ courses/models/__init__.py
@@ -1,0 +1,2 @@
+from .course import Course
+from .wagtail import *
