+++ courses/models/tariffstats.py
@@ -1,0 +1,73 @@
+from .tariff import Tariff
+from .utils import fallback_to, enums, separate_unavail_reason
+
+class TariffStatistics:
+
+    def __init__(self, tariff_data, display_language):
+        self.display_language = display_language
+        self.tariffs = []
+
+        if tariff_data:
+            self.aggregation = tariff_data.get('aggregation')
+            self.number_of_students = fallback_to(tariff_data.get('number_of_students'), 0)
+            if tariff_data.get('tariffs'):
+                for tariff in tariff_data.get('tariffs'):
+                    self.tariffs.append(Tariff(tariff, self.display_language))
+            self.tariffs.reverse()
+
+            subject_data = tariff_data.get('subject')
+            if subject_data:
+                self.subject_code = subject_data.get('code')
+                self.subject_english = subject_data.get('english_label')
+                self.subject_welsh = subject_data.get('welsh_label')
+
+            unavailable_data = fallback_to(tariff_data.get('unavailable'), {})
+            self.unavailable_code = unavailable_data.get('code')
+            self.unavailable_reason = fallback_to(unavailable_data.get('reason'), '')
+            self.unavailable_reason_english = fallback_to(unavailable_data.get('reason_english'), '')
+            self.unavailable_reason_welsh = fallback_to(unavailable_data.get('reason_welsh'), '')
+            self.unavailable_find_out_more_english = fallback_to(unavailable_data.get('find_out_more_english'), '')
+            self.unavailable_find_out_more_welsh = fallback_to(unavailable_data.get('find_out_more_welsh'), '')
+            self.unavailable_url_english = fallback_to(unavailable_data.get('url_english'), '')
+            self.unavailable_url_welsh = fallback_to(unavailable_data.get('url_welsh'), '')
+
+    def show_stats(self):
+        return self.tariffs
+
+    def display_subject_name(self):
+        if self.display_language == enums.languages.ENGLISH:
+            return self.subject_english if self.subject_english else self.subject_welsh
+        return self.subject_welsh if self.subject_welsh else self.subject_english
+
+    def display_unavailable_info(self):
+        unavailable = {}
+
+        if self.unavailable_reason:
+            unavailable["reason"] = self.unavailable_reason
+        else:
+            if self.display_language == enums.languages.ENGLISH:
+                unavailable["reason"] = self.unavailable_reason_english if self.unavailable_reason_english \
+                    else self.unavailable_reason_welsh
+            else:
+                unavailable[
+                    "reason"] = self.unavailable_reason_welsh if self.unavailable_reason_welsh else self.unavailable_reason_english
+
+        if self.display_language == enums.languages.ENGLISH:
+            unavailable[
+                "find_out_more"] = self.unavailable_find_out_more_english if self.unavailable_find_out_more_english \
+                else self.unavailable_find_out_more_welsh
+        else:
+            unavailable[
+                "find_out_more"] = self.unavailable_find_out_more_welsh if self.unavailable_find_out_more_welsh else self.unavailable_find_out_more_english
+
+        if self.display_language == enums.languages.ENGLISH:
+            unavailable["url"] = self.unavailable_url_english if self.unavailable_url_english \
+                else self.unavailable_url_welsh
+        else:
+            unavailable[
+                "url"] = self.unavailable_url_welsh if self.unavailable_url_welsh else self.unavailable_url_english
+
+        unavailable["reason_heading"], unavailable["reason_body"] = separate_unavail_reason(unavailable["reason"])
+
+        return unavailable
+
